src/kitchen/requirements.txt
paho-mqtt
prometheus_client
prometheus_flask_exporter 
pydantic
redis